Add files via upload
This is the implementation of the Runyankore grammar engine as a Protege plugin. It contains the plugin itself, the XML annotation files, as well as the test ontologies. The plugin should be placed in the "plugins" folder of your Protege distribution. You can generate Runyankore text by loading one of the test ontologies, and clicking the "Runyankore>Verbalizer" submenu under the "Tools" menu